Next release 1.1.3

+++ app/src/main/java/com/sports/unity/messages/controller/activity/GroupInfoFragment.java
@@ -214,6 +214,10 @@ public class GroupInfoFragment extends Fragment implements ActivityCompat.OnRequ
         }
         String partCount = getResources().getString(R.string.participant_count);
         partCount = String.format(partCount, groupParticipants.usersInGroup.size());
+        if(groupJID.startsWith(Constants.DISCUSS_JID)){
+            partCount = getResources().getString(R.string.discuss_participant_count);
+            partCount = (groupParticipants.usersInGroup.size()-1)+" Participants";
+        }
         groupCount.setText(partCount);
 
         participantsList = (ListView) view.findViewById(R.id.participants_list);
@@ -563,14 +567,22 @@ public class GroupInfoFragment extends Fragment implements ActivityCompat.OnRequ
     };
 
     private void exitAndDeleteGroup() {
-        boolean success = exitGroup();
-        if (success) {
-            deleteGroup();
+        if (groupJID.startsWith(Constants.DISCUSS_JID)) {
+            showInDeterminateProgress("Please wait...");
+            UserProfileHandler.getInstance().submitGroupExit(getActivity(), groupJID, LISTENER_KEY);
+        } else {
+            boolean success = exitGroup();
+            if (success) {
+                deleteGroup();
+            }
         }
     }
 
     private void deleteGroup() {
         SportsUnityDBHelper.getInstance(getContext()).deleteGroup(chatID);
+        if (groupJID.startsWith(Constants.DISCUSS_JID)) {
+            SportsUnityDBHelper.getInstance(getContext()).deleteDiscussionDetail(groupJID);
+        }
         NotificationHandler.getInstance(getActivity().getApplicationContext()).clearNotificationMessages(String.valueOf(chatID));
         Intent I = new Intent(getActivity(), MainActivity.class);
         I.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP | Intent.FLAG_ACTIVITY_SINGLE_TOP);
@@ -689,6 +701,24 @@ public class GroupInfoFragment extends Fragment implements ActivityCompat.OnRequ
                             Toast.makeText(getActivity(), R.string.message_submit_failed, Toast.LENGTH_SHORT).show();
                         }
 
+                        dismissInDeterminateProgress();
+                    }
+
+                });
+            } else if (requestTag.equals(UserProfileHandler.SUBMIT_GROUP_LEAVING_REQUEST_TAG)) {
+                final Boolean success = (Boolean) content;
+                getActivity().runOnUiThread(new Runnable() {
+
+                    @Override
+                    public void run() {
+                        if (success) {
+                            boolean success = exitGroup();
+                            if (success) {
+                                deleteGroup();
+                            }
+                        } else {
+                            Toast.makeText(getActivity(), R.string.message_exit_failed, Toast.LENGTH_SHORT).show();
+                        }
                         dismissInDeterminateProgress();
                     }
